+++ database/migrations/2020_02_04_143319_foreign_key_users.php
@@ -14,7 +14,7 @@ class ForeignKeyUsers extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->unsignedInteger('role_id')->default(1);
+            $table->unsignedBigInteger('role_id')->default(1);
             $table->foreign('role_id')->references('id')->on('roles'); 
         });
     }
